Prevent unnecessary canceling of CA requests
https://github.com/neovim/neovim/commit/4bfdd1ee9dee15c4b066412335f8cafbc3208f60
The above commit (4bfdd1e) in neovim logs verbosely if a request no
longer exists when trying to cancel, which revealed that we were calling
the cancel function for a completed request errneously.

End result is we should no longer see lines like:

[ERROR][2025-03-26 16:12:59] ...m/lsp/client.lua:623	"LSP[lua_ls]"
"Cannot find request with id 27 whilst attempting to cancel"

being spammed in the LspLog.